requirements.txt
requests>=2.28.0
pyyaml>=6.0
click>=8.1.0
colorama>=0.4.6
flask>=2.3.0
psutil>=5.9.0
werkzeug>=2.3.0
